fix(show-last): drop unnecessary i64 cast for clippy on CI
Linux CI clippy flags casting civil_to_unix_secs (already i64) to i64.

+++ src-tauri/src/agents/transcript_full.rs
@@ -320,7 +320,7 @@ pub fn parse_cursor_wall_clock_label(label: &str) -> Option<u64> {
     }
     let offset_secs = parse_utc_offset_label(tz)?;
     // Civil time in that offset → UTC unix.
-    let utc_secs = civil_to_unix_secs(year, month, day, hour, minute, 0)? as i64 - offset_secs;
+    let utc_secs = civil_to_unix_secs(year, month, day, hour, minute, 0)? - offset_secs;
     if utc_secs < 0 {
         return None;
     }
